Finishes redux update
