+++ app/templates/_react/_redux/_package.json
@@ -1,6 +1,6 @@
 {
   "scripts": {
-    "start-iso": "ISO_TEST=True NODE_ENV=production babel-node server.js"
+    "start": "$(npm bin)/babel-node server.js"
   },
   "dependencies": {
     "async-props": "^0.3.2",
@@ -8,6 +8,7 @@
     "babel-cli": "^6.6.5",
     "babel-preset-stage-0": "^6.5.0",
     "babel-preset-stage-2": "^6.5.0",
+    "babel-runtime": "^6.11.6",
     "ejs": "^2.4.1",
     "eslint": "^2.13.0",
     "eslint-plugin-react": "^2.7.1",
@@ -26,9 +27,13 @@
     "redux-promise": "^0.5.3",
     "redux-thunk": "^2.1.0",
     "redux-wait": "0.0.3",
+    "superagent": "^2.2.0",
     "webpack-stream": "^3.1.0",
     "window-or-global": "^1.0.1",
     "xhr-request": "^1.0.1"
+  },
+  "devDependencies": {
+    "babel-plugin-transform-runtime": "^6.15.0"
   }
 }
