+++ app/templates/_react/_redux/babelrc
@@ -1,3 +1,4 @@
 {
-  "presets": ["es2015", "stage-0", "stage-2", "react"]
+  "presets": ["es2015", "stage-0", "stage-2", "react"],
+  "plugins": ["transform-runtime"]
 }
