+++ app/templates/_react/_redux/src/app/Home/Home.jsx
@@ -1,4 +1,5 @@
 import React        from 'react';
+import request      from 'superagent';
 import {connect}    from 'react-redux';
 import {setTitle}   from '../actions/home';
