--- app/templates/_react/_redux/src/app/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>your app</title>
-    <base href="/">
-    <!-- vendor:css -->
-    <!-- endinject -->
-    <!-- head:css -->
-    <!-- endinject -->
-    <meta name="viewport" content="initial-scale=1, maximum-scale=1">
-    <!-- inject:googleanalytics -->
-    <!-- endinject -->
-    <link href="./index.css" rel="stylesheet" type="text/css">
-    <link href="images/icon.png" rel="icon" type="image/png">
-  </head>
-  <body>
-    <div id="content">
-      <%- reactContent %>
-    </div>
-    <!-- vendor:js -->
-    <!-- endinject -->
-    <script src="./main.js"></script>
-  </body>
-</html>
