--- app/templates/_react/_redux/src/lib/alt.jsx
@@ -1,2 +1,0 @@
-import Alt from 'alt';
-export default new Alt();
